Split mpsc channel into smaller modules
